Port initial implementations from script
It's apparent there's a lot of coupling between functionality in the
various functions for the shell script implementation. This is leading
me to think that we can do a 1:1 port but then do a lot of refactoring
that needs to take place regardless.

In particular, some functions do setup with specific state on the
filesystem and other require that said state be in place. With that in
mind, it might make sense to simply port the primary logic into
something monolithic and abstract things out from there.

There's also some _new_ functionality in here regarding caching the
targets. The goal is to pull the latest version of the targets down and
stash them someplace; serializing them if they are already present on
running the tool. I may only get a little bit of that functionality in
place here.