Other minor tweaks
I think at this stage it would be helpful to start with a clean slate
and consider what the ideal API would be. We can see a lot of the
general cases here from this partial port. I am imagining things such as

`InstallCache`   - for both cabal and GHC installs
`GHCTargetCache` - for mapping available (GHC) targets to their
                   respective URLs
`GHCTargetCache` - for mapping available (cabal) targets to their
                   respective URLs
`Target`         - for housing functionality around downloading and
                   installation re: both GHC and cabal
`G`              - alternative for `Target`; houses the targets, the
                   target cahce, and the install cache as well as
                   providing a handle around installing, removing,
                   listing, etc.