Port initial implementations from script (#38)
* Port initial implementations from script

It's apparent there's a lot of coupling between functionality in the
various functions for the shell script implementation. This is leading
me to think that we can do a 1:1 port but then do a lot of refactoring
that needs to take place regardless.

In particular, some functions do setup with specific state on the
filesystem and other require that said state be in place. With that in
mind, it might make sense to simply port the primary logic into
something monolithic and abstract things out from there.

There's also some _new_ functionality in here regarding caching the
targets. The goal is to pull the latest version of the targets down and
stash them someplace; serializing them if they are already present on
running the tool. I may only get a little bit of that functionality in
place here.

* Download and stash targets

* Grab machine specific targets

* Notes on next moves

* Remove stub test

* Other minor tweaks

I think at this stage it would be helpful to start with a clean slate
and consider what the ideal API would be. We can see a lot of the
general cases here from this partial port. I am imagining things such as

`InstallCache`   - for both cabal and GHC installs
`GHCTargetCache` - for mapping available (GHC) targets to their
                   respective URLs
`GHCTargetCache` - for mapping available (cabal) targets to their
                   respective URLs
`Target`         - for housing functionality around downloading and
                   installation re: both GHC and cabal
`G`              - alternative for `Target`; houses the targets, the
                   target cahce, and the install cache as well as
                   providing a handle around installing, removing,
                   listing, etc.

* Reflect ring removal in cargo lockfile

* Break up `system_and_machine`

Also removes `os_to_target`.

* Depend on tempfile crate

* Break up lib.rs into modules

* Begin integrating g_lib into g_cli

* Appease clippy

* Initial (untested) install functionality

As noted in the code, it may make sense to make `download` and `install`
a single fn to avoid callers of the library forgetting to do one before
the other. This way there is less burden on the client to 'do the right
thing'.

* Stubs more legacy functionality

* Download step in G rather than `old` module

* Add types for Machine and System

* Dump old.rs file

* Fix CI